Move the test parsing helper code.
