+++ tests/attribute_tests.rs
@@ -1,12 +1,12 @@
 // Copyright (c) ZeroC, Inc.
 
-pub mod helpers;
+pub mod test_helpers;
 
 mod attributes {
 
     mod slice_api {
 
-        use crate::helpers::parsing_helpers::*;
+        use crate::test_helpers::*;
         use slice::diagnostics::{Error, ErrorKind, Warning, WarningKind};
         use slice::grammar::*;
         use test_case::test_case;
@@ -541,7 +541,7 @@ mod attributes {
 
     mod generalized_api {
 
-        use crate::helpers::parsing_helpers::parse_for_ast;
+        use crate::test_helpers::parse_for_ast;
         use slice::compile_from_strings;
         use slice::grammar::*;
         use test_case::test_case;
