+++ tests/classes/encoding.rs
@@ -2,7 +2,7 @@
 
 mod slice2 {
 
-    use crate::helpers::parsing_helpers::*;
+    use crate::test_helpers::*;
     use slice::diagnostics::{Error, ErrorKind};
     use slice::grammar::Encoding;
