+++ tests/exceptions/tags.rs
@@ -1,6 +1,6 @@
 // Copyright (c) ZeroC, Inc.
 
-use crate::helpers::parsing_helpers::parse_for_ast;
+use crate::test_helpers::parse_for_ast;
 use slice::grammar::*;
 
 #[test]
