--- tests/helpers/mod.rs
@@ -1,4 +1,0 @@
-// Copyright (c) ZeroC, Inc.
-
-// TODO: move the parsing helpers directly into this file.
-pub mod parsing_helpers;
