+++ tests/identifier_tests.rs
@@ -1,8 +1,8 @@
 // Copyright (c) ZeroC, Inc.
 
-pub mod helpers;
+pub mod test_helpers;
 
-use helpers::parsing_helpers::parse_for_ast;
+use crate::test_helpers::parse_for_ast;
 use slice::grammar::{Enum, Exception, Interface, Struct};
 
 #[test]
