+++ tests/interfaces/mod.rs
@@ -4,7 +4,7 @@ mod encoding;
 mod inheritance;
 mod operations;
 
-use crate::helpers::parsing_helpers::*;
+use crate::test_helpers::*;
 use slice::diagnostics::{Error, ErrorKind};
 use slice::grammar::*;
