+++ tests/primitives/encoding.rs
@@ -2,7 +2,7 @@
 
 mod slice1 {
 
-    use crate::helpers::parsing_helpers::*;
+    use crate::test_helpers::*;
     use slice::diagnostics::{Error, ErrorKind};
     use slice::grammar::Encoding;
     use test_case::test_case;
@@ -76,7 +76,7 @@ mod slice1 {
 
 mod slice2 {
 
-    use crate::helpers::parsing_helpers::*;
+    use crate::test_helpers::*;
     use slice::diagnostics::{Error, ErrorKind};
     use slice::grammar::Encoding;
     use test_case::test_case;
