+++ tests/primitives/mod.rs
@@ -2,7 +2,7 @@
 
 mod encoding;
 
-use crate::helpers::parsing_helpers::parse_for_ast;
+use crate::test_helpers::parse_for_ast;
 use slice::grammar::*;
 use test_case::test_case;
